# backend/src/schemas/openapi.yaml
openapi: 3.1.0
info:
  title: Graphfolio Backend API
  description: Backend API for Graphfolio - Company network visualization
  version: 1.0.0
servers:
- url: https://api.tinboker.com
  description: Production server
- url: https://staging-api.tinboker.com
  description: Staging server
- url: https://dev-api.tinboker.com
  description: Development server
- url: http://localhost:8000
  description: Local development
paths:
  /api/graphs:
    get:
      tags:
      - graphs
      summary: Get Sorted Graphs
      description: 'Get sorted graphs list


        Query params:

        - sort_by: Sort field (concept_id, created_at, updated_at)'
      operationId: get_sorted_graphs_api_graphs_get
      parameters:
      - name: sort_by
        in: query
        required: false
        schema:
          type: string
          default: concept_id
          title: Sort By
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                title: Response Get Sorted Graphs Api Graphs Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - graphs
      summary: Create Graph
      description: 'Create new graph


        Creates a graph with nodes and edges. Stock information will be fetched for
        nodes.'
      operationId: create_graph_api_graphs_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GraphCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/graphs/{graph_id}:
    get:
      tags:
      - graphs
      summary: Get Graph By Id
      description: 'Get graph by ID


        Returns complete graph with nodes and edges'
      operationId: get_graph_by_id_api_graphs__graph_id__get
      parameters:
      - name: graph_id
        in: path
        required: true
        schema:
          type: string
          description: Graph ID
          title: Graph Id
        description: Graph ID
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GraphData'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - graphs
      summary: Delete Graph
      description: 'Delete graph


        Deletes graph and all its nodes and edges'
      operationId: delete_graph_api_graphs__graph_id__delete
      parameters:
      - name: graph_id
        in: path
        required: true
        schema:
          type: string
          description: Graph ID
          title: Graph Id
        description: Graph ID
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/graphs/{graph_id}/nodes/{node_id}:
    put:
      tags:
      - graphs
      summary: Modify Node
      description: 'Modify node in graph


        Updates node position and/or data'
      operationId: modify_node_api_graphs__graph_id__nodes__node_id__put
      parameters:
      - name: graph_id
        in: path
        required: true
        schema:
          type: string
          description: Graph ID
          title: Graph Id
        description: Graph ID
      - name: node_id
        in: path
        required: true
        schema:
          type: string
          description: Node ID
          title: Node Id
        description: Node ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NodeUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - graphs
      summary: Delete Node
      description: 'Delete node from graph


        Deletes node and all connected edges'
      operationId: delete_node_api_graphs__graph_id__nodes__node_id__delete
      parameters:
      - name: graph_id
        in: path
        required: true
        schema:
          type: string
          description: Graph ID
          title: Graph Id
        description: Graph ID
      - name: node_id
        in: path
        required: true
        schema:
          type: string
          description: Node ID
          title: Node Id
        description: Node ID
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/graphs/{graph_id}/edges/{edge_id}:
    put:
      tags:
      - graphs
      summary: Modify Edge
      description: 'Modify edge in graph


        Updates edge data'
      operationId: modify_edge_api_graphs__graph_id__edges__edge_id__put
      parameters:
      - name: graph_id
        in: path
        required: true
        schema:
          type: string
          description: Graph ID
          title: Graph Id
        description: Graph ID
      - name: edge_id
        in: path
        required: true
        schema:
          type: string
          description: Edge ID
          title: Edge Id
        description: Edge ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EdgeUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - graphs
      summary: Delete Edge
      description: Delete edge from graph
      operationId: delete_edge_api_graphs__graph_id__edges__edge_id__delete
      parameters:
      - name: graph_id
        in: path
        required: true
        schema:
          type: string
          description: Graph ID
          title: Graph Id
        description: Graph ID
      - name: edge_id
        in: path
        required: true
        schema:
          type: string
          description: Edge ID
          title: Edge Id
        description: Edge ID
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/news:
    get:
      tags:
      - news
      summary: Get Sorted News
      description: 'Get sorted news list


        Query params:

        - sort_by: Sort field (date, created_at, updated_at, title)


        CDN Cache: 10 minutes'
      operationId: get_sorted_news_api_news_get
      parameters:
      - name: sort_by
        in: query
        required: false
        schema:
          type: string
          description: Sort field
          default: date
          title: Sort By
        description: Sort field
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StockEvent'
                title: Response Get Sorted News Api News Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/news/{news_id}:
    get:
      tags:
      - news
      summary: Get News By Id
      description: 'Get news by ID


        Returns complete news event with related tickers


        CDN Cache: 10 minutes'
      operationId: get_news_by_id_api_news__news_id__get
      parameters:
      - name: news_id
        in: path
        required: true
        schema:
          type: string
          description: News ID
          title: News Id
        description: News ID
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockEvent'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/news/fetch/{ticker}:
    post:
      tags:
      - news
      summary: Fetch News From Massive
      description: "Fetch news from Massive API for a ticker and save to database\n\
        \nArgs:\n    ticker: Stock ticker symbol (e.g., \"NVDA\", \"AAPL\")\n    limit:\
        \ Maximum number of articles to fetch (default: 10)\n    \nReturns:\n    Dict\
        \ with count of news articles fetched and their IDs"
      operationId: fetch_news_from_massive_api_news_fetch__ticker__post
      parameters:
      - name: ticker
        in: path
        required: true
        schema:
          type: string
          description: Stock ticker symbol
          title: Ticker
        description: Stock ticker symbol
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          description: Maximum number of articles to fetch
          default: 10
          title: Limit
        description: Maximum number of articles to fetch
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Fetch News From Massive Api News Fetch  Ticker  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    Edge:
      properties:
        id:
          type: string
          title: Id
          description: Unique edge identifier
        source:
          type: string
          title: Source
          description: Source node ID (ticker)
        target:
          type: string
          title: Target
          description: Target node ID (ticker)
        label:
          type: string
          title: Label
          description: Display label for the edge
        data:
          allOf:
          - $ref: '#/components/schemas/EdgeData'
          description: Edge data payload
      type: object
      required:
      - id
      - source
      - target
      - label
      - data
      title: Edge
      description: Graph edge representing relationship between stocks
    EdgeCreate:
      properties:
        id:
          type: string
          title: Id
          description: Unique edge identifier
        source:
          type: string
          title: Source
          description: Source node ID (ticker)
        target:
          type: string
          title: Target
          description: Target node ID (ticker)
        label:
          type: string
          title: Label
          description: Display label for the edge
        category:
          type: string
          title: Category
          description: Edge category
      type: object
      required:
      - id
      - source
      - target
      - label
      - category
      title: EdgeCreate
      description: Edge creation model
    EdgeData:
      properties:
        category:
          type: string
          title: Category
          description: Edge category (e.g., 'aiChips', 'automation', 'components')
      type: object
      required:
      - category
      title: EdgeData
      description: Data payload for graph edge
    EdgeUpdate:
      properties:
        source:
          anyOf:
          - type: string
          - type: 'null'
          title: Source
          description: Source node ID (ticker)
        target:
          anyOf:
          - type: string
          - type: 'null'
          title: Target
          description: Target node ID (ticker)
        label:
          anyOf:
          - type: string
          - type: 'null'
          title: Label
          description: Display label for the edge
        category:
          anyOf:
          - type: string
          - type: 'null'
          title: Category
          description: Edge category
      type: object
      title: EdgeUpdate
      description: Edge update model - all fields optional
...
    GraphCreate:
      properties:
        conceptId:
          type: string
          enum:
          - robotics
          - ai
          - energy
          title: Conceptid
          description: Concept identifier
        nodes:
          items:
            $ref: '#/components/schemas/NodeCreate'
          type: array
          title: Nodes
          description: List of nodes
        edges:
          items:
            $ref: '#/components/schemas/EdgeCreate'
          type: array
          title: Edges
          description: List of edges
      type: object
      required:
      - conceptId
      - nodes
      - edges
      title: GraphCreate
      description: Graph creation model
    GraphData:
      properties:
        nodes:
          items:
            $ref: '#/components/schemas/Node'
          type: array
          title: Nodes
          description: List of graph nodes
        edges:
          items:
            $ref: '#/components/schemas/Edge'
          type: array
          title: Edges
          description: List of graph edges
      type: object
      required:
      - nodes
      - edges
      title: GraphData
      description: Complete graph structure with nodes and edges
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Node:
      properties:
        id:
          type: string
          title: Id
          description: Unique node identifier (ticker)
        type:
          type: string
          title: Type
          description: Node type (e.g., 'stock')
        data:
          allOf:
          - $ref: '#/components/schemas/NodeData'
          description: Node data payload
        position:
          allOf:
          - $ref: '#/components/schemas/Position'
          description: Node position in graph
      type: object
      required:
      - id
      - type
      - data
      - position
      title: Node
      description: Graph node representing a stock
    NodeCreate:
      properties:
        id:
          type: string
          title: Id
          description: Unique node identifier (ticker)
        type:
          type: string
          title: Type
          description: Node type
          default: stock
        label:
          type: string
          title: Label
          description: Display label for the node
        ticker:
          type: string
          title: Ticker
          description: Stock ticker symbol
        marketCapTier:
          type: string
          title: Marketcaptier
          description: Market capitalization tier
        positionX:
          type: number
          title: Positionx
          description: X coordinate
        positionY:
          type: number
          title: Positiony
          description: Y coordinate
      type: object
      required:
      - id
      - label
      - ticker
      - marketCapTier
      - positionX
      - positionY
      title: NodeCreate
      description: Node creation model
    NodeData:
      properties:
        label:
          type: string
          title: Label
          description: Display label for the node
        ticker:
          type: string
          title: Ticker
          description: Stock ticker symbol
        marketCapTier:
          type: string
          title: Marketcaptier
          description: Market capitalization tier (large/medium/small)
      type: object
      required:
      - label
      - ticker
      - marketCapTier
      title: NodeData
      description: Data payload for graph node
    NodeUpdate:
      properties:
        label:
          anyOf:
          - type: string
          - type: 'null'
          title: Label
          description: Display label for the node
        ticker:
          anyOf:
          - type: string
          - type: 'null'
          title: Ticker
          description: Stock ticker symbol
        marketCapTier:
          anyOf:
          - type: string
          - type: 'null'
          title: Marketcaptier
          description: Market capitalization tier
        positionX:
          anyOf:
          - type: number
          - type: 'null'
          title: Positionx
          description: X coordinate
        positionY:
          anyOf:
          - type: number
          - type: 'null'
          title: Positiony
          description: Y coordinate
      type: object
      title: NodeUpdate
      description: Node update model - all fields optional
...
    Position:
      properties:
        x:
          type: number
          title: X
          description: X coordinate
        y:
          type: number
          title: Y
          description: Y coordinate
      type: object
      required:
      - x
      - y
      title: Position
      description: 2D position coordinates
...
    StockEvent:
      properties:
        id:
          type: string
          title: Id
          description: Unique event identifier
        type:
          type: string
          enum:
          - earnings
          - conference
          - news
          - dividend
          title: Type
          description: Event type
        date:
          type: integer
          title: Date
          description: Event date (Unix timestamp in milliseconds)
        title:
          type: string
          title: Title
          description: Event title
        description:
          type: string
          title: Description
          description: Event description
        content:
          anyOf:
          - type: string
          - type: 'null'
          title: Content
          description: Event content
        relatedTickers:
          items:
            type: string
          type: array
          title: Relatedtickers
          description: List of related stock tickers
      type: object
      required:
      - id
      - type
      - date
      - title
      - description
      - relatedTickers
      title: StockEvent
      description: Stock-related event (earnings, news, etc.)
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer